Reorganized the structure after reading the typer docs
