+++ photomise/utilities/logging.py
